Conflicts:
	KmsBase.sln
	Templates/Driver_KMDF/Driver_KMDF.vcxproj
	_DocDev/Procedures/Build.txt

+++ KmsCopy/KmsCopy.cpp
@@ -169,7 +169,7 @@ int Execute(FILE * aFile, const char * aSrcRoot, const char * aDstRoot)
 			char lDstFile[1024];
 			char lSrcFile[1024];
 
-			switch (sscanf_s(lLine, " %[^ \n\r\t] %[^ \n\r\t]", lSrcFile, sizeof(lSrcFile) / sizeof(lSrcFile[0]), lDstFile, sizeof(lDstFile) / sizeof(lDstFile[0])))
+			switch (sscanf_s(lLine, " %[^ \n\r\t] %[^ \n\r\t]", lSrcFile, static_cast<int>(sizeof(lSrcFile) / sizeof(lSrcFile[0])), lDstFile, static_cast<int>(sizeof(lDstFile) / sizeof(lDstFile[0]))))
 			{
 			case EOF: break; // Empty line
 
@@ -206,7 +206,7 @@ int Execute(FILE * aFile, const char * aSrcRoot, const char * aDstRoot)
 			break;
 
 		default : // Folder
-			switch (sscanf_s(lLine, "%[^ \n\r\t] %[^ \n\r\t]", lSrcFolder, sizeof(lSrcFolder) / sizeof(lSrcFolder[0]), lDstFolder, sizeof(lDstFolder) / sizeof(lDstFolder[0])))
+			switch (sscanf_s(lLine, "%[^ \n\r\t] %[^ \n\r\t]", lSrcFolder, static_cast<int>(sizeof(lSrcFolder) / sizeof(lSrcFolder[0])), lDstFolder, static_cast<int>(sizeof(lDstFolder) / sizeof(lDstFolder[0]))))
 			{
 			case 1:
 				strcpy_s(lDstFolder, lSrcFolder);
